Refactor: move all decisions regarding a single ACL entry down the chain
I.e., move some early conditionalizing from pcmk__unpack_acl down to
__xml_acl_parse_entry.  This is so as to prepare for parsing "acl_group"
along with already established handling of "acl_target" -- which of
these is getting handled needs to be propagated down the call stack,
and with this first step of descent, it makes sense for the caller to
parse this information on its own, since the whole XML subtree is passed
down anyway.  We already take the opportunity to generalize the context
of ACLs unpacking from a single scalar (username) into a vector of
strings (resolved per-user groups will follow soon).