Fix: libcrmcommon: Remove --output-meta-refresh=.
This command line option is only used in crm_mon, which also has a
--interval= option that does the same thing.  The latter works just fine
because it is translated into --output-meta-refresh.  However, the
former does not work at all because crm_mon has no way of knowing
whether it was used.  Thus, --output-meta-refresh ends up always getting
passed with the --interval default and whatever the user specified on
the command line will be ignored.

This is fixed by simply eliminating that command line option.  Doing so
forces the command line tools to set up the refresh header, so I've also
added pcmk__html_add_header that makes the job a little easier.