Log: libcib: improve messages for CIB sign-on
Don't print CIB sign-on errors to stderr, or log sign-on messages above info
level, because some callers might want to print more contextual messages or not
consider it an error.  Also, printing directly to stderr is not appropriate for
daemons, and doesn't play well with the new output model.

Instead, callers should check the return code and print or log a message if
desired. Nearly all already did anyway.