Fix: scheduler: ignore colocation constraints with 0 score
A positive score means to prefer the same node, and a negative score means to
prefer a different node, so presumably a zero score should have no effect.

Previously, the 0 score would sometimes cause the colocation itself to be
ignored, but indirect preferences from resources colocated with the resource in
the colocation would still be incorporated. For example:

   colocate A with B score 100
   colocate B with C score 0

would cause C to incorporate preferences from A.

Colocations with 0 score are still added to the data set, however. This means
they are still relevant to crm_resource --recursive and --constraints (which
is perhaps questionable).