Refactor: scheduler: make showing node weights more efficient
The scores_log_level global in libpacemaker wasn't really needed since it was
constant and not particularly worthwhile to abstract. dump_node_scores() used
its log severity argument as a boolean, so make that explicit, and rename the
function according to current style guidelines.

Use the new pcmk__log_else() macro to skip iterating through all nodes
(and again recursively for all child resources) when we're not tracing,
since we wouldn't do anything in that case anyway.