Log: scheduler: simplify logging resource info
print_resource() was mostly useless. It always used trace level logging, but
the user would have to know to trace the resource's print() implementation to
see it. It is now incorporated into the one place it was useful (print_node()),
and all other uses have been replaced with pe_rsc_trace().