Fix: libcrmcommon: Make sure to free the --output-as description.
This is done by putting it into pcmk__common_args_t which ensures it
will be freed when g_option_context_free is called.  Also, add
pcmk__free_arg_context as a wrapper around that function.  This makes
sure context is valid first and gives us a place to put additional
cleanups later.