Log: scheduler: include execution time when logging action failure
The CIB keeps track of a resource's action history necessary to reconstruct its
current state. When the scheduler processes failure history, it logs the failure,
regardless of whether it is the first scheduler run since the failure happened
or not. Users are often confused by this and think the failure is recurring.
So, include the time when the failure result happened in the log message.