Feature: tools: Add --include= and --exclude= to crm_mon.
These options give a single place to specify which portions of the
crm_mon output should actually be printed to the screen.  Each can be
specified multiple times, and each invocation can include a
comma-separated list of sections.  Giving an unrecognized section will
cause a list of valid sections to be printed.

These options can also be mixed with the older options like --operations
and --failcounts since they are all just different ways of manipulating
the "show" bitfield.