Refactor: libraries: prevent internal code from using deprecated API
Define a new PCMK__NO_COMPAT symbol in crm_internal.h, which is used
by all internal code. Headers can declare deprecated APIs guarded by
that symbol, so they are available to external API callers but not
to internal callers.

Some deprecated symbols are used internally and thus are not yet
guarded. Where possible, internal code should be updated to use the
new forms and the old forms moved to the guarded sections.