Build: finalize: restore buildability in the face of obsolete ftime(3)
This is a second phase of what's started with 4b8b84cce1 (the message
lacked "finalize" tag).

We proceed to drop ftime(3) uses and surrounding old cruft altogether,
but compatibility/contingency plan is provided:

    in case of CLOCK_MONOTONIC detected in time.h positively but
    choking for whatever reason in the actual build or even in run-time,
    please reconfigure with CPPFLAGS=-DPCMK_TIME_EMERGENCY_CGT,
    which will cause no time diffing (remember, optional) will
    get compiled in at all;
    alternatively, you can shortcut any checking and refrain from
    any time period measurement with something (not future-proof!)
    like:

      env \
        ac_cv_have_decl_CLOCK_MONOTONIC=no \
        ./configure

Note that what was unspotted originally (for reliance on the original
code's sanity) is that neither systemd nor nagios support is supposed
to work entirely correctly without any measuring of time periods
available, so until the deeper overhaul of those parts, we will just
refrain from those classes of resources to have a support in the build
when "clock_gettime(CLOCK_MONOTONIC, ...)" appears unsupported.