Feature: ANSI codes with access-render-2.xsl also configurable per user
The same overriding approach is also applied for developmental use from
codebase checkout.  The trick is to hook into libxslt's resource loader
subsystem, and be hence able to override, on-the-fly, the URI to be
loaded, per some prior detections of what's suitable the most for
one such in particular -- already known xsl:include'd
access-render.cfg.xsl.