Fix: tools: Correct the crm_mon man page.
The notes and time specifications were not being output correctly.
Previously, I was using asterisks around the section names to set them
apart in the man page.  This served to turn those names into proper man
page sections, while as the same time looking reasonably good in the
help output.  However, this asterisk trick does not work with older
versions of help2man, such as is used in the epel7 mock chroot (and
therefore presumably, on RHEL7 builders).

Instead, add the headers with the man page include file.  However, this
means the man page will have both the groff header and the plain text
header surrounded by .PP lines.  tools/fix-manpages is a sed script
that removes those blocks.

This documentation now looks reasonable in both the man page and in the
help output.