Fix: tools: Move the cluster-options list closing hack.
Instead of doing it inside the cluster-options message function, do it
inside print_cluster_summary.  The former is going to move into the
pengine library, which shouldn't know about such things like the context
it is called in.  Leave that knowledge in crm_mon.

Unfortunately, this means certain crm_mon printing functions need to
know about the output format again.