Fix: tools: Do not use mon_ops in the node message.
The node message functions are going to be moving into pengine.  They
use mon_ops in a variety of places, but that's a crm_mon specific
concept.  pengine doesn't know anything about mon_ops and we should keep
it that way.

Most of the things it uses mon_ops for can simply just be turned into
extra boolean arguments to the functions.