Log: all: downgrade some xpath search errors to LOG_NEVER
If get_xpath_object() doesn't find anything, it logs error messages (including
dumping the entire XML input) at a given log severity. For many callers, the
search result is optional; these would use LOG_TRACE to be as quiet as possible
about it. However that was still annoying when tracing the relevant code.
Now, use LOG_NEVER in such cases, so nothing is logged for a search with no
hits when the search target is optional.