Low: controller: mark shutdown-locked resources in resource history
When a graph action indicates that the resource should be shutdown-locked
to its node, remember the shutdown lock time in active_op_t so we can remember
that when the result comes back. When the result does come back, add
"shutdown-lock" to its lrm_resource entry in the CIB status section -- as
the timestamp if it's a successful stop or a probe finding the resource
inactive, or as 0 to clear the lock for any other operation.