Build: doc: improve upload targets
The main purpose of this is to create new targets for "books", "books-upload",
and "books-clean" separated out from the www and clean-local targets, so the
books can be generated and uploaded separately from all the other
documentation.

The opportunity is also taken to rename variables for clarity and make
similar new targets for the remaining deprecated ascii/html documentation.