Fix: Mark formatted output functions with a special attr.
This attribute contains type information for the varargs used by
formatted output functions.  This information can be used by other tools
to determine whether these functions are being called correctly.

This is a non-standard attribute - it isn't supported or understood by
any compiler.  It is hidden behind a macro, and that macro only
evaluates to anything if a special cpp define is present.  That define
is set when compiling with the gcc-python-plugin.  Under normal
compilation, there should be no difference.