Fix: libstonithd, tools: Fix arguments to stonith-event.
First, the second argument should be a gboolean instead of an int.  Use
is_set when calling to make sure that is the case.  Second,
stonith_event_xml should pull all varargs off using va_arg.  This
prevents bugs in the future should the number of args change.  Third,
print_opts should be an unsigned int instead of an int.  It's a bit
field.