Fix: libcrmcommon: request should be a gchar *, not a char *.
It is constructed by processing argv with g_strjoinv, which returns a
gchar *.  This also requires making processed_args a gchar *, which it
was in some places but not others.  Finishing that requires changing the
return type of pcmk__cmdline_preproc.

Finally, call the correct freeing functions for request and
processed_args.