Low: scheduler: respect shutdown locks when placing active resources
Use new pe_resource_t members to indicate that a resource is locked to a
particular node.

For active resources (i.e. in the transition where the node is scheduled for
shutdown), these are connected by checking each lockable resource for whether
it is running on a single clean node that is shutting down.

When applying constraints, place -INFINITY location constraints for locked
resources on all nodes other than the lock node.

(Inactive resources -- i.e. in later transitions after the node is shut down --
are not yet locked.)