Refactor: libpacemaker: Add higher level stonith functions.
This is the beginnings of moving the useful stuff out of the command
line tools and into libraries.  The plan is that over time, the tools
will basically just process command line arguments and call into the
libraries.  All the code that does the hard work will be in a library.

There will be two layers to the library.  The pcmk__ versions are for
internal use only.  If they do output, they'll take a pcmk__output_t
that has already been set up.  The command line tools will use this
layer.

The other layer builds on top of that.  These are the pcmk_ versions and
are for external users.  If they do output, they'll take an xmlNodePtr
(or some similar well-structured destination) and write their results to
that.  Behind the scenes, they will set up the pcmk__output_t and pass
that to the pcmk__ version.

For some functions, the pcmk_ version just calls to the pcmk__ one
without doing anything else.  These exist just to keep the external API
complete.