doc: skip outdated part of example log
The function outputting the transition statistics is not
process_pe_message since commit 4d5fe6bf (present in 2.0.2), but
pcmk__log_transition_summary, so that part is different in the actual
log now. The first part of the log line (up to "notice") is subject to
change anyway, so just omit it altogether for simplicity.