Doc: all: remove translations
Publican is a dead project, and it is getting difficult to generate the
documentation on newer distros. As a first step to moving away from Publican,
remove all translations, which were generated via Publican (they have not been
used for years now anyway).

If we want to add translations again in the future, we can come up with a new
model.