Fix: libpacemaker: Timeouts should be in ms, given as an unsigned int.
Eventually, we'd like timeouts throughout public APIs to be
milliseconds.  This starts that by accepting ms for the new API
functions and converting them to seconds before passing them in to
stonith commands.  Should the stonith commands later be converted to
take ms, it would be easy to modify these calls to adapt.

This also applies to the tolerance parameter.