Build: libpacemaker: Disable building the public API.
There are other refactorings that need to take place before this API is
good enough to be exposed to outside users.  In particular, stonith_t
and stonith_key_value_t are likely to be changed.  I don't want to
release a public API using them that then needs immediate changes.

So for the moment, the public API is disabled unless the right ifdef is
provided.