Fix: libpe_status: Add a resource-header output message.
The current resource-history message is serving two different purposes.
First, it creates a new list with a header in print_rsc_history.
Second, it adds an item to an existing list in print_node_history.
These two need to be formatted slightly differently, so a new message
seems like the easiest way to do it.

This only affects the text and console outputs.

Also, put single quotes around the date in the resource history strings.