Fix: tools: Re-enable CGI output from crm_mon.
The CGI header was not being written out because "false" was being
passed to the finish function.  That was being passed because we didn't
want the HTML to be printed out without the refresh header.  The fix is
just to s/false/true, and change the order so the extra header is added
first.