Fix: tools: Maintain a list of extra HTML headers.
This list of headers then gets added to the head node when finish is
called.  This fixes a problem where finish needed to be called first (to
make sure the head node was created), but calling it first meant you had
to choose between getting CGI headers and getting the extra headers.