Fix: libpacemaker: Fix handling of some operation return values.
The list_agents and query stonith operations return a negative value on
error or a positive value corresponding to the number of matches.  The
libpacemaker API needs to return the appropriate error if a negative
value was received and otherwise return pcmk_rc_ok.  Callers of this
library probably do not care about the number of matches.

This fixes most of the failures in CTS.