Low: controller: allow CRM_OP_LRM_DELETE to clear CIB only
Normally, CRM_OP_LRM_DELETE is relayed to the affected node's controller, which
clears the resource from the executor and CIB as well the its own bookkeeping.

Now, we want to be able to use it to clear shutdown locks for nodes that are
down. Let it take a new "mode" attribute, and if it is "cib", clear the
resource from the CIB locally without relaying the operation or doing anything
else.