Fix: libpe_status: Use the same message for headers and history.
resource-history gets used in two places - sometimes as the header of a
list, and sometimes as the body of a list.  Because colons are appended
to the text of a list header, this results in extra colons in the output
sometimes.

This patch adds an extra argument to resource-history to specify whether
it is a list header or list body.  This is only needed on text/console
output.

Note also that an end_list call is removed from print_rsc_history.  This
call was only present to close out the list that resource-history had
previously been unconditionally created.