Fix: cts: Add tests for zero or more than one date_expression.
First, convert the existing test for no date_expression (which was never
really a case a user could get into anyway) into a test that has more
than one date_expression.  Second, add a test that has an expression
that is not a date_expression.  This is valid XML that the user could
actually throw at us.  Changing those two things requires adding the
same XML blocks in many places, which is the bulk of the changes to this
file.

Then, change some strings in the CTS test cases that have changed over
time.

Finally, change the error message in crm_rule.c to reflect the
possibility that the found rule could have zero date_expressions or more
than one.