Fix: libpe_status: Close the list that pe__resource_history_xml makes.
I previously thought there was no need to deal with the as_header
parameter to this XML output function.  However, in print_node_history
in crm_mon, the branch that prints out history instead of operations
will open a new list.  This list never gets closed anywhere which means
any other XML nodes will end up being children of that list, which is
not correct.

So just use as_header to determine whether or not the list should
immediately be closed or not.