Fix: tests: Add initial tests for pe_cron_range_satisfied.
There's lots more about this function that could be tested.  I'm not
doing a very thorough job testing all the possiblilities that can be
given in a date_spec, but they are just variations that use the same
underlying code for comparisons.

The basic layout of unit tests is:  Under each directory in lib/, we
will have a tests subdirectory.  Within each of those directories, there
will further be subdirectories for each source code file.  And then
within each of those directories, there will be a file for each function
that is being tested.

Each directory under lib/*/tests/ should have the same name as the
source file it tests.  Each file under lib/*/tests/* should have the
same name as the function it tests.  Remember to add the compiled unit
test to .gitignore, too.