Feature: tools: Allow crm_rule to check some date_specs.
Previously, crm_rule was only able to check rules that contained a
single date_expression, and that could not contain a date_spec.  This
patch opens up crm_rule to check date_specs that contain years= and do
not contain moon=.

This is done entirely by changing the xpath expression that searches for
date_expression nodes and then by checking for satisfied/unsatisfied in
the return value.  No additional work is needed in actually evaluating
the expressions.

Note that date_specs without years= do not ever expire and therefore
can be handled with the current behavior.