Refactor: libcrmcommon: Change return value of pcmk__split_range.
It should return one of the standard pacemaker return codes, just like
all new code should.  It returns pcmk_rc_unknown_format if the string is
empty, NULL, or just contains the separator.  It returns pcmk_rc_ok in
all other cases.  One thing to watch out for on pcmk_rc_ok - if both
start and end are NULL, this means the string contained a single number,
not a range.