Fix: cts: Correct the test for too many date_exps in crm_rule.
In the call to crm_rule in cts-cli.in, I passed an incorrect name for
the rule to check.  This meant the rule was not found and therefore the
wrong test was being performed.  I'm now passing the correct name for
the rule with too many date_expressions.  This requires adding the right
return code definition to cts-cli.in as well.