Low: executor: record correct last run and last rc change times
This is a regression introduced by 4b8b84cc, which replaced ftime() calls
with calls to clock_gettime(CLOCK_MONOTONIC, ...).

However the last run and last rc change times were passed back to the client as
epoch timestamps, which would end up in the CIB resource operation history (and
crm_mon output).

The last rc change time was never used to calculate an interval, which means it
never really needed to use ftime() to begin with -- a time_t set to time(NULL)
would have been sufficient, which is what is now used.

For the last run, which is also used to calculate an interval (queue time and
execution time), we now store both the monotonic timestamp (for interval
calculation) and an equivalent epoch time_t (for return to the client).