Doc: all: create sphinx infrastructure
Publican is a dead project, and it is getting difficult to generate the
documentation on newer distros, so we are going to transition all our
"books" from asciidoc + Publican to reStructuredText + sphinx.

This commit creates the infrastructure for using sphinx. This uses a
"sphinx" subdirectory of doc that is not built by default, so the Publican
version of the books will still be what is installed.