Refactor: tools, libpe_status: Create new cluster summary messages.
These messages sit on top of the existing cluster-* messages, calling
all of them to construct the complete cluster summary.

Doing all of this as a message means the block in print_cluster_summary
that used output_format to decide whether to close a list can go away.
There's now simply a different message for the HTML case that does the
list closing.  That then means output_format can stop being passed to
all the print_*_status functions because it isn't needed anymore.

Also, note the new block in print_status to try to ensure that a blank
line gets printed out after the summary, but only if the summary gets
printed.