Fix: tools: Apply various fixes to the crm_diff man page.
First, remove the asterisks from around the "Examples" header and put a
colon on the end of the line.  This will make it into a section header
in both the --help output and the man page even on RHEL7.  See
4f5207a28a991abfbea3e0fe54177ac4a8ff8959 for more information.

Second, put # signs on each line to indicate they are command lines.
Also add an extra newline to the end of the first cibadmin --query
example so the second is on its own line.  They were being run together
in the man page output.

Third, the cibadmin --patch example also needs a -x.