Fix: xml: Add a new version of the tags schema.
First, there was an extra space in the name of the obj_ref's id
attribute.  That can't have been helping anything.

Second, change the type of that attribute from an IDREF to just text.
Nothing with tags was validating because there's no ID type in the nodes
XML.  And changing that seems more disruptive as could require changing
existing IDs to use valid characters (no leading numbers, for instance).

The other potential fix for this would be to apply the following to
xml/nodes-3.0.rng:

     <element name="nodes">
       <zeroOrMore>
         <element name="node">
-          <attribute name="id"><text/></attribute>
+          <attribute name="id"><data type="ID"/></attribute>
           <attribute name="uname"><text/></attribute>
           <optional>
             <attribute name="type">

However, this is impossible to do.  The contents of the id attribute
here are defined by the cluster layer.  These contents are not
restricted to only those values that are valid for an ID.  That is, the
id attribute is allowed to contain characters that the ID type
disallows.

I've updated the cts-cli output to match the new version of the schema.