Feature: libstonithd: Report an error getting history in output.
Currently, there is no way to differentiate between the fencing history
being empty and an error occurring when getting the history.  Both are
simply NULL pointers.  This patch adds a way to take the return code
from getting the history and adding that as an attribute to the XML
output.  This value is not reported in any other output formats at the
moment.