Feature: scheduler: add priority-fencing-delay cluster option
This feature addresses the relevant topics and implements the ideas
brought up from:

https://github.com/ClusterLabs/fence-agents/pull/308

This commit adds priority-fencing-delay option (just the option, not the
feature itself).

Enforce specified delay for the fencings that are targeting the lost
nodes with the highest total resource priority in case we don't
have the majority of the nodes in our cluster partition, so that
the more significant nodes potentially win any fencing match,
which is especially meaningful under split-brain of 2-node
cluster. A promoted resource instance takes the base priority + 1
on calculation if the base priority is not 0. If all the nodes
have equal priority, then any pcmk_delay_base/max configured for
the corresponding fencing resources will be applied. Otherwise as
long as it's set, even if to 0, it takes precedence over any
configured pcmk_delay_base/max. By default, priority fencing
delay is disabled.