Fix: tools: One more blank line related fix for crm_mon.
The previous patch using pcmk_rc_no_output was close, but not completely
correct.  It failed to account for the situation where block A has
output, block B was requested but has no output, and block C has output.
In this case, rc will be set to pcmk_rc_no_output from the call to block
B, resulting in no blank line between A and C.

Instead, what we need to do is keep track of if any block before the
previous one had output, only setting rc to pcmk_rc_ok and never back to
pcmk_rc_no_output.