Fix: scheduler: don't incorporate dependencies' colocations multiple times
When placing a resource, the scheduler incorporates the preferences of
other resources that will be colocated with the first resource.

Previously, if the dependent resource was a group, the first resource would
incorporate the preferences of each group member in turn. The problem is that
each group member also incorporates the preferences of each group member after
it. Thus the same constraint scores were being added multiple times, distorting
the desired effect.