Test: scheduler: update regression tests for colocation scoring changes
I believe but am not 100% sure these changes are reasonable. There is a
somewhat unexpected effect on clone instance allocation, but the end result is
the same. The scoring changes imply there must be some situations where the
behavior would change, but it does not appear those would be problematic.